requirements.txt
# Smart Notes Generator - Dependencies
# Generated: 2026-01-25
# Python 3.8+

# Core Video Processing
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Image Processing & Analysis
scikit-image>=0.21.0
Pillow>=10.0.0

# Data Manipulation & Analysis
pandas>=2.0.0
numpy>=1.24.0

# Scientific Computing
scipy>=1.10.0

# Utilities
pathlib2>=2.3.7
tqdm>=4.65.0

# Optional: GPU Acceleration (uncomment if CUDA available)
# opencv-python-headless>=4.8.0

# Machine Learning
xgboost>=1.7.0

# Notes Generation Dependencies
paddlepaddle>=2.6.0  # Core PaddlePaddle framework (required for PaddleOCR)
paddleocr>=2.7.0
google-generativeai>=0.8.0
openai-whisper>=20231117  # Standard Whisper (better Windows compatibility)
markdown>=3.5.0
# weasyprint>=60.0  # Optional: For PDF generation - install separately if needed

# Fix for grpcio dependency conflicts
grpcio>=1.60.0,<2.0.0
grpcio-status>=1.60.0,<2.0.0

# Development Dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
# jupyter>=1.0.0
